+++ src/main/java/com/jesua/registration/dto/UserDto.java
@@ -13,5 +13,5 @@ public class UserDto {
     private String password;
     private String role;
     private Boolean active;
-    private int projectId;
+    private long projectId;
 }
